+++ tests/e2e/api/src/models/products/index.ts
@@ -3,3 +3,5 @@ export * from './shared';
 export * from './simple-product';
 export * from './grouped-product';
 export * from './external-product';
+export * from './variation';
+export * from './variable-product';